Move MatWrapper to vision module head
